pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xbom"
version = "0.1.0"
description = "Generate unified SBOM, AI-BOM, CBOM, and API-BOM in CycloneDX 1.6 JSON"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "Aviral" }]
keywords = ["sbom", "aibom", "cbom", "api-bom", "cyclonedx", "security", "supply-chain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "cyclonedx-python-lib>=8.0,<9.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
apibom = [
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-java>=0.23",
    "tree-sitter-go>=0.23",
    "tree-sitter-typescript>=0.23",
]
agent = [
    "strands-agents>=0.1",
    "strands-agents-tools>=0.1",
]
codegraph = [
    "falkordb>=1.1,<2.0",
    "falkordb-code-graph>=0.1",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]
all = ["xbom[apibom,agent,codegraph,dev]"]

[project.scripts]
xbom = "xbom.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"xbom" = ["**/*.yaml"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
